Add JSDoc for custom hooks, helper methods, and API related code
