+++ jobs/Frontend/movie-search/src/configs/appRoutes.ts
@@ -1,6 +1,18 @@
-const enum AppRoutes {
+/**
+ * An enumeration representing the application route paths.
+ */
+export const enum AppRoutes {
+  /**
+   * The route path for the search functionality.
+   */
   Search = '/search',
+  /**
+   * The route path for the movie section.
+   */
   Movie = '/movie',
+  /**
+   * The route path for the "page not found" error handler.
+   */
   PageNotFound = '/page-not-found'
 }
