+++ jobs/Frontend/movie-search/src/features/api/searchMovie.tsx
@@ -1,11 +1,19 @@
 import authenticatedApiClient from '../auth/authenticatedApiClient';
-import { ApiData } from '../common/models/ApiData';
 import { Movie } from '../movies/models/Movie';
+import { ApiData } from './models/ApiData';
 
 interface MovieData extends Omit<ApiData, 'results'> {
   results: Movie[];
 }
 
-export async function searchMovie(query: string) {
+/**
+ * Function that searches for movies based on a query string.
+ *
+ * @param query - The search query string to use for the API request.
+ * @returns An object containing movie data, including an array of {@link Movie} objects in the 'results' property.
+ *
+ * @throws Throws an error if the API request fails.
+ */
+export async function searchMovies(query: string) {
   return await authenticatedApiClient.get<MovieData>(`/search/movie?query=${query}`);
 }
