--- jobs/Frontend/movie-search/src/features/common/models/ApiData.ts
@@ -1,6 +1,0 @@
-export interface ApiData {
-  page: number;
-  results: [];
-  total_pages: number;
-  total_results: number;
-}
